1.4.6

+++ fast_sql/fastsql/sql.py
@@ -6,15 +6,15 @@ import re
 import sys
 import threading
 import shutil
+from concurrent.futures import ThreadPoolExecutor, as_completed
+
 import pandas as pd
 from tqdm import tqdm
-from queue import Queue
-from fast_sql.utils.exception import DB_Exceptions
-from concurrent.futures import ThreadPoolExecutor
+
 from fast_sql.utils.common import DB_Pool, collection_error, Sqlalchemy_Pool
 
 
-class Read_sql:
+class ReadSql:
     def __init__(
         self,
         sql,
@@ -28,7 +28,7 @@ class Read_sql:
         auto_order=False,
     ):
         if con is not None:
-            self.db_pool = Sqlalchemy_Pool(con, num=thread_num + 6, encoding=encoding)
+            self.db_pool = Sqlalchemy_Pool(con, num=thread_num + 1, encoding=encoding)
             self.driver = self.db_pool.driver
         self.sql = sql.strip().replace("\n", " ")
         self.thread_num = thread_num
@@ -39,7 +39,6 @@ class Read_sql:
         self.ordering = None
         self.show_progress = show_progress
         self.progress = None
-        self.queue = Queue()
         self.count = None
         self.tqdm = None
         self.chunksize = chunksize
@@ -47,6 +46,9 @@ class Read_sql:
         self.desc = desc
         self.columns_upper = columns_upper
         self.auto_order = auto_order
+        self.consumer_count = 0
+        self.pool = ThreadPoolExecutor(max_workers=self.thread_num)
+        self.task_list = []
 
     def read_sql(self, **kwargs):
         con = self.db_pool.get_db()
@@ -59,8 +61,8 @@ class Read_sql:
                 result.columns = [i.upper() for i in result.columns.tolist()]
             self.tqdm_update(self.count)
         else:
-            task_list, pool = self.start_thread_read()
-            self.get_thread_result(task_list)
+            self.start_thread_read()
+            self.get_thread_result()
             result = self.result.pop(0)
             for i in range(len(self.result)):
                 result = result.append(self.result.pop(0), ignore_index=True)
@@ -71,7 +73,11 @@ class Read_sql:
                 result.drop("NO", axis=1, inplace=True)
 
         if self.auto_order:
-            result.sort_values([i.strip() for i in self.ordering], inplace=True,ascending=self.sort_type)
+            result.sort_values(
+                [i.strip() for i in self.ordering],
+                inplace=True,
+                ascending=self.sort_type,
+            )
         self.db_pool.close_db(con)
         self.result.clear()
         return result
@@ -81,19 +87,18 @@ class Read_sql:
         if self.show_progress:
             if count:
                 tqdm.update(count)
-
             elif tqdm.n + self.chunksize < self.count:
                 tqdm.update(self.chunksize)
 
             else:
                 tqdm.update(self.count - tqdm.n)
 
-    def tqdm_init(self, count, weight, mode=None):
+    def tqdm_init(self, count, weight, mode=None, desc=None):
         if self.show_progress:
             if mode is None:
-                self.tqdm = tqdm(total=count, desc=self.desc, ncols=weight)
+                self.tqdm = tqdm(total=count, desc=desc or self.desc, ncols=weight)
             else:
-                self.tqdm_w = tqdm(total=count, desc=self.desc, ncols=weight)
+                self.tqdm_w = tqdm(total=count, desc=desc or self.desc, ncols=weight)
 
     @collection_error
     def get_sql_query(self, st, en, *args, **kwargs):
@@ -105,7 +110,7 @@ class Read_sql:
             """
         else:
             sql_1 = f"""
-                 select * from ({self.sql}) as t limit {st}{','+str(en)}
+                 select * from ({self.sql}) as t limit {st}{',' + str(en)}
             """
         con = self.db_pool.get_db()
         query = pd.read_sql(sql_1, con, **self.pd_params)
@@ -122,11 +127,8 @@ class Read_sql:
     def start_thread_read(
         self,
     ):
-        pool = ThreadPoolExecutor(max_workers=self.thread_num)
-        task_list = [
-            pool.submit(self.get_sql_query, st, en) for st, en in self.avg_list
-        ]
-        return task_list, pool
+        for st, en in self.avg_list:
+            self.task_list.append(self.pool.submit(self.get_sql_query, st, en))
 
     def verify_sql(self):
         con = self.db_pool.get_db()
@@ -134,13 +136,21 @@ class Read_sql:
         # self.chunksize = self.count // self.thread_num // 2
         # self.tqdm_init(self.count, desc='Read the scheduler', weight=85)
         if self.auto_order:
-            buf = re.search("(?<=order[\s+]by)(.*)(?=limit)",self.sql,re.I)
+            buf = re.search("(?<=order[\s+]by)(.*)(?=limit)", self.sql, re.I)
             if buf:
                 buf = buf.group()
-                self.sort_type = False if 'desc' in buf.lower() else True
-                buf = buf.replace('desc','').replace('asc','').replace('DESC','').replace('ASC','')
-                self.sql = 'select ' + re.search("(?<=select)(.*)(?=limit)",self.sql,re.I).group()
-                self.ordering = buf.split(',')
+                self.sort_type = False if "desc" in buf.lower() else True
+                buf = (
+                    buf.replace("desc", "")
+                    .replace("asc", "")
+                    .replace("DESC", "")
+                    .replace("ASC", "")
+                )
+                self.sql = (
+                    "select "
+                    + re.search("(?<=select)(.*)(?=limit)", self.sql, re.I).group()
+                )
+                self.ordering = buf.split(",")
 
         if self.count < 2000:
             return None
@@ -161,15 +171,13 @@ class Read_sql:
         return avg_list
 
     def get_query_count(self, con, sql):
-        # _sql = "select count(*) " + re.search(r"from\s+.*", sql, re.I).group()
         _sql = f"select count(*) from ({sql}) t"
         count = pd.read_sql(_sql, con, **self.pd_params).iloc[0, 0]
-        # assert count > 0, DB_Exceptions('DB_EMPTY:')
         return count
 
-    def get_thread_result(self, task_list):
-        for job in task_list:
-            job.result()
+    def get_thread_result(self):
+        for task in as_completed(self.task_list):
+            task.result()
 
     def __del__(self):
         if self.show_progress:
@@ -178,9 +186,9 @@ class Read_sql:
         sys.stdout.flush()
 
 
-class to_csv(Read_sql):
+class ToCsv(ReadSql):
     def __init__(self, *args, path_or_buf=None, **kwargs):
-        super(to_csv, self).__init__(*args, **kwargs)
+        super(ToCsv, self).__init__(*args, **kwargs)
         self.path = path_or_buf
         self.args = None
         self.kwargs = None
@@ -214,9 +222,9 @@ class to_csv(Read_sql):
             df = pd.read_sql(self.write_csv_header(), con)
             # df = next(pd.read_sql(self.sql, con, chunksize=1)).iloc[1:, ]
             df.to_csv(*args, **kwargs)
-            task_list, pool = self.start_thread_read()
+            self.start_thread_read()
             self.db_pool.close_db(con)
-            self.get_thread_result(task_list)
+            self.get_thread_result()
 
         self.db_pool.close_db(con)
         return "finish"
@@ -239,7 +247,7 @@ class to_csv(Read_sql):
         self.lock.release()
 
 
-class to_sql(Read_sql):
+class ToSql(ReadSql):
     def __init__(self, *args, **kwargs):
         to_db = kwargs.pop("to_db")
         kwargs.update(con=kwargs.pop("from_db"))
@@ -250,19 +258,23 @@ class to_sql(Read_sql):
         self.file_path = kwargs.pop("file_path")
         self.delete_cache = kwargs.pop("delete_cache")
         self.save_path = kwargs.pop("save_path", None)
-        self.thread_w = kwargs.pop("thread_w", 3)
+        self.thread_w = kwargs.pop("thread_w", 5)
         self.delete_sql = kwargs.pop("delete_sql", None)
         self.dir_path = None
         self.task_count = None
         self.execute_count = 0
         self.data_processing = kwargs.pop("data_processing", None)
-        self.extra_param = kwargs.pop('extra_param',None)
+        self.extra_param = kwargs.pop("extra_param", None)
+        self.write_desc = kwargs.pop("write_desc", "write the scheduler")
+        self.df = kwargs.pop("df", None)
         super().__init__(*args, **kwargs)
         if to_db is not None:
             self.to_db = DB_Pool(to_db, num=8, encoding=kwargs.get("encoding"))
             self.write_driver = self.to_db.driver
         self.lock_b = threading.Lock()
         self.execute_pool = None
+        self.w_pool = ThreadPoolExecutor(max_workers=self.thread_w)
+        self.task_list = []
 
     def delete_table(self):
         sql = self.sql.replace(
@@ -293,44 +305,37 @@ class to_sql(Read_sql):
             df.to_pickle(file_path)
             if self.mode in ("wr", "rw"):
                 self.task_count = 1
-                self.tqdm_init(self.task_count, weight=85, mode=True)
-                self.queue.put(file_path)
-                self.write_db()
+                self.tqdm_init(
+                    self.task_count, weight=85, mode=True, desc=self.write_desc
+                )
+                self.insert_db(file_path)
             else:
                 self.tqdm_update(count=self.task_count)
-
             self.db_pool.close_db(con)
 
         else:
-            task_list, pool = self.start_thread_read()
             self.task_count = len(self.avg_list)
-            if self.mode in ("wr", "rw"):
-                self.tqdm_init(self.task_count, weight=85, mode=True)
-                _pool = ThreadPoolExecutor(max_workers=self.thread_w)
-                p = [pool.submit(self.write_db) for i in range(5)]
-                self.get_thread_result(p)
-            self.get_thread_result(task_list)
+            self.tqdm_init(self.task_count, weight=85, mode=True, desc=self.write_desc)
+            self.start_thread_read()
+            self.get_thread_result()
 
     def rsync_db(self, *args, **kwargs):
-
         self.pd_params = kwargs
-
         self.sql = self.sql.strip().replace("\n", " ")
-
         if self.to_table is None:
             self.to_table = re.search(r"from\s+(\S+)", self.sql, re.I).group(1)
-
         if self.if_exists == "delete" and self.mode != "r":
             self.delete_table()
-
         if self.mode == "w":
             return self.write()
-
+        if self.df is not None:
+            self.count = 1
+            self.task_count = 1
+            self.tqdm_init(1, weight=85, mode="w")
+            return self.insert_db(df=self.df)
         self.avg_list = self.verify_sql()
         self.tqdm_init(self.count, weight=85)
-
         self.decision()
-
         return "finish"
 
     def write(self):
@@ -338,20 +343,19 @@ class to_sql(Read_sql):
             os.path.join(self.file_path, i) for i in os.listdir(self.file_path)
         ]
         self.task_count = len(file_list)
-        self.tqdm_w = tqdm(total=self.task_count, desc=self.desc, ncols=80)
-        T = ThreadPoolExecutor(max_workers=self.thread_w)
-        put_list = [self.queue.put(path) for path in file_list]
-        task_list = [T.submit(self.write_db) for i in put_list]
-        T.shutdown(wait=True)
-        return "finish"
+        self.tqdm_w = tqdm(total=self.task_count, desc=self.write_desc, ncols=80)
+        self.task_list.extend(
+            [self.w_pool.submit(self.insert_db, i) for i in file_list]
+        )
+        self.get_thread_result()
 
     def save_query(self, df_value):
         file_path = os.path.join(self.dir_path, f"{uuid.uuid1()}.pkl")
         if self.driver == "oracle":
             df_value.drop(df_value.columns[0], axis=1, inplace=True)
         df_value.to_pickle(file_path)
-        self.queue.put(file_path)
         self.tqdm_update()
+        self.task_list.append(self.w_pool.submit(self.insert_db, file_path))
 
     def get_sys_guid_col(self, columns, con):
         col_type = pd.read_sql(
@@ -372,9 +376,7 @@ class to_sql(Read_sql):
         columns,
     ):
         insert_col = ",".join(columns)
-        # con = self.to_db.get_db()
         if self.write_driver == "oracle":
-            # columns = self.get_sys_guid_col(columns, con)
             sql_col = ":" + ",:".join([str(i) for i in range(1, len(columns) + 1)])
             sql = f"insert into {self.to_table}({insert_col}) values({sql_col})"
             return sql
@@ -386,27 +388,15 @@ class to_sql(Read_sql):
             return sql
 
     @collection_error
-    def write_db(self):
-        while True:
-            self.lock_b.acquire()
-            if self.execute_count < self.task_count:
-                # df = pd.read_pickle(file_path)
-                self.execute_count += 1
-                self.lock_b.release()
-                path = self.queue.get()
-                self.insert_db(path)
-            else:
-                # self.tqdm_w.update(self.task_count - self.tqdm_w.n)
-                self.lock_b.release()
-                break
-
-    def insert_db(self, path):
-        df = pd.read_pickle(path)
+    def insert_db(self, path=None, df=None):
+        df = pd.read_pickle(path) if path else df
         if self.columns_upper:
             df.columns = [i.upper() for i in df.columns]
         if self.data_processing is not None:
-            if 'extra_param' in inspect.signature(self.data_processing).parameters:
-                df = self.data_processing(df,extra_param=copy.deepcopy(self.extra_param))
+            if "extra_param" in inspect.signature(self.data_processing).parameters:
+                df = self.data_processing(
+                    df, extra_param=copy.deepcopy(self.extra_param)
+                )
             else:
                 df = self.data_processing(df)
 
@@ -450,7 +440,7 @@ class to_sql(Read_sql):
     def __del__(self):
 
         if self.delete_cache:
-            if self.mode in ("wr", "rw") and self.save_path is None:
+            if self.mode in ("wr", "rw") and self.save_path is None and self.df is None:
                 shutil.rmtree(self.file_path)
         if self.show_progress:
             if self.tqdm is not None:
